% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{objects}
\alias{objects}
\title{Object and Link Type Registration}
\description{
Functions for registering object types (mappings from logical
concepts to physical tables) and link types (relationships between objects).
}
