% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{ont_action_summary}
\alias{ont_action_summary}
\title{Action Summary}
\usage{
ont_action_summary(action_type_id = NULL, since = NULL, con = NULL)
}
\arguments{
\item{action_type_id}{Optional filter by action type.}

\item{since}{Optional datetime to filter actions after.}

\item{con}{Optional DBI connection.}
}
\value{
A tibble with summary statistics.
}
\description{
Returns summary statistics for actions.
}
