% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{ont_get_cohort_members}
\alias{ont_get_cohort_members}
\title{Get Cohort Members}
\usage{
ont_get_cohort_members(cohort_id, con = NULL)
}
\arguments{
\item{cohort_id}{Character. The cohort ID.}

\item{con}{A DBI connection. If \code{NULL}, uses the active connection.}
}
\value{
A character vector of object keys.
}
\description{
Retrieves the members of a cohort (evaluates SQL for dynamic cohorts).
}
