% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concepts.R
\name{ont_get_concept}
\alias{ont_get_concept}
\title{Get Concept Metadata}
\usage{
ont_get_concept(concept_id, con = NULL)
}
\arguments{
\item{concept_id}{Character. The concept ID.}

\item{con}{A DBI connection.}
}
\value{
A single-row tibble.
}
\description{
Get Concept Metadata
}
\keyword{internal}
