% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{ont_get_concept_inheritance}
\alias{ont_get_concept_inheritance}
\title{Get Concept Inheritance}
\usage{
ont_get_concept_inheritance(concept_id, con = NULL)
}
\arguments{
\item{concept_id}{Character. The concept ID.}

\item{con}{A DBI connection. If \code{NULL}, uses the active connection.}
}
\value{
A tibble of templates this concept inherits from.
}
\description{
Returns the template(s) that a concept inherits from.
}
\examples{
\dontrun{
ont_get_concept_inheritance("unemployed_ireland")
}

}
