% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{ont_get_connection}
\alias{ont_get_connection}
\title{Get Active Connection}
\usage{
ont_get_connection()
}
\value{
A DBI connection object.
}
\description{
Returns the active database connection. Throws an error if not connected.
}
\keyword{internal}
