% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{ont_get_score_components}
\alias{ont_get_score_components}
\title{Get Score Components}
\usage{
ont_get_score_components(score_id, con = NULL)
}
\arguments{
\item{score_id}{The score ID.}

\item{con}{Optional DBI connection.}
}
\value{
A tibble of components.
}
\description{
Retrieves the components of a composite score.
}
