% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{ont_get_template}
\alias{ont_get_template}
\title{Get Template Details}
\usage{
ont_get_template(template_id, con = NULL)
}
\arguments{
\item{template_id}{Character. The template ID.}

\item{con}{A DBI connection. If \code{NULL}, uses the active connection.}
}
\value{
A list with template details including parsed parameters.
}
\description{
Retrieves information about a specific template.
}
\examples{
\dontrun{
template <- ont_get_template("ilo_unemployed")
template$parameters
}

}
