% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{ont_get_template_variants}
\alias{ont_get_template_variants}
\title{Get Template Variants}
\usage{
ont_get_template_variants(template_id, con = NULL)
}
\arguments{
\item{template_id}{Character. The template ID.}

\item{con}{A DBI connection. If \code{NULL}, uses the active connection.}
}
\value{
A tibble of concepts that inherit from the template.
}
\description{
Lists all concepts that inherit from a specific template.
}
\examples{
\dontrun{
ont_get_template_variants("ilo_unemployed")
}

}
