% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concepts.R
\name{ont_get_version}
\alias{ont_get_version}
\title{Get Concept Version}
\usage{
ont_get_version(concept_id, scope, version, con = NULL)
}
\arguments{
\item{concept_id}{Character. The concept ID.}

\item{scope}{Character. The scope.}

\item{version}{Integer. The version number.}

\item{con}{A DBI connection.}
}
\value{
A single-row tibble with version details.
}
\description{
Get Concept Version
}
\keyword{internal}
