% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ont_import_json}
\alias{ont_import_json}
\title{Import Ontology from JSON}
\usage{
ont_import_json(path, overwrite = FALSE, con = NULL)
}
\arguments{
\item{path}{Character. Path to JSON file.}

\item{overwrite}{Logical. Overwrite existing definitions? Default FALSE.}

\item{con}{A DBI connection. If \code{NULL}, uses the active connection.}
}
\value{
Invisibly returns TRUE.
}
\description{
Imports ontology definitions from a JSON file.
}
