% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{ont_list_apps}
\alias{ont_list_apps}
\title{Check Available Shiny Apps}
\usage{
ont_list_apps()
}
\value{
A character vector of app names.
}
\description{
Lists all available Shiny applications in the ontologyR package.
}
\examples{
\dontrun{
ont_list_apps()
}

}
