% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{ont_list_datasets}
\alias{ont_list_datasets}
\title{List Datasets}
\usage{
ont_list_datasets(dataset_type = NULL, object_type = NULL, con = NULL)
}
\arguments{
\item{dataset_type}{Optional filter by dataset type.}

\item{object_type}{Optional filter by object type.}

\item{con}{Optional DBI connection.}
}
\value{
A tibble of datasets.
}
\description{
Returns a list of all registered datasets.
}
