% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{ont_list_layers}
\alias{ont_list_layers}
\title{List Spatial Layers}
\usage{
ont_list_layers(object_type = NULL, con = NULL)
}
\arguments{
\item{object_type}{Optional filter by object type.}

\item{con}{Optional DBI connection.}
}
\value{
A tibble of layers.
}
\description{
Lists all defined spatial layers.
}
