% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{ont_list_regions}
\alias{ont_list_regions}
\title{List Spatial Regions}
\usage{
ont_list_regions(con = NULL)
}
\arguments{
\item{con}{Optional DBI connection.}
}
\value{
A tibble of regions.
}
\description{
Lists all defined spatial regions.
}
