% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{ont_list_scores}
\alias{ont_list_scores}
\title{List Scores}
\usage{
ont_list_scores(object_type = NULL, con = NULL)
}
\arguments{
\item{object_type}{Optional filter by object type.}

\item{con}{Optional DBI connection.}
}
\value{
A tibble of scores with component counts.
}
\description{
Lists all defined composite scores.
}
