% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{ont_list_templates}
\alias{ont_list_templates}
\title{List All Templates}
\usage{
ont_list_templates(object_type = NULL, source_standard = NULL, con = NULL)
}
\arguments{
\item{object_type}{Character. Filter by object type (optional).}

\item{source_standard}{Character. Filter by source standard (optional).}

\item{con}{A DBI connection. If \code{NULL}, uses the active connection.}
}
\value{
A tibble of templates.
}
\description{
Returns a tibble of all defined templates.
}
\examples{
\dontrun{
ont_list_templates()
ont_list_templates(source_standard = "ILO")
}

}
