% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{ont_list_transforms}
\alias{ont_list_transforms}
\title{List Transforms}
\usage{
ont_list_transforms(
  output_dataset_id = NULL,
  transform_type = NULL,
  con = NULL
)
}
\arguments{
\item{output_dataset_id}{Optional filter by output dataset.}

\item{transform_type}{Optional filter by type.}

\item{con}{Optional DBI connection.}
}
\value{
A tibble of transforms.
}
\description{
Returns all defined transforms.
}
