% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{ont_query_linked}
\alias{ont_query_linked}
\title{Query Linked Objects}
\usage{
ont_query_linked(from_type, from_ids, link_type, as_of = NULL, con = NULL)
}
\arguments{
\item{from_type}{Character. Source object type.}

\item{from_ids}{Vector of source object primary keys.}

\item{link_type}{Character. The link type to traverse.}

\item{as_of}{Date or timestamp. For temporal links, return links valid at
this point in time. If \code{NULL}, returns all links.}

\item{con}{A DBI connection. If \code{NULL}, uses the active connection.}
}
\value{
A lazy dplyr tbl with joined target objects.
}
\description{
Starting from objects of one type, traverse a link to get related objects.
}
