% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{ont_remove_score_component}
\alias{ont_remove_score_component}
\title{Remove Score Component}
\usage{
ont_remove_score_component(score_id, component_id, con = NULL)
}
\arguments{
\item{score_id}{The score ID.}

\item{component_id}{The component to remove.}

\item{con}{Optional DBI connection.}
}
\value{
TRUE if successful.
}
\description{
Removes a component from a score.
}
