% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{ont_status}
\alias{ont_status}
\title{Get Connection Status}
\usage{
ont_status()
}
\value{
A list with connection details, or \code{NULL} if not connected.
}
\description{
Returns information about the current ontology database connection.
}
