% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{ont_writeback_action}
\alias{ont_writeback_action}
\title{Write Action Results to Table}
\usage{
ont_writeback_action(action_id, action_type, object_key, params, actor, con)
}
\arguments{
\item{action_id}{The action ID.}

\item{action_type}{The action type record.}

\item{object_key}{The object key.}

\item{params}{The action parameters.}

\item{actor}{The user who executed the action.}

\item{con}{DBI connection.}
}
\value{
The result of the writeback.
}
\description{
Internal function to write action results to the configured writeback table.
}
\keyword{internal}
