% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_provenance}
\alias{print_provenance}
\title{Print Method for Ontology Provenance}
\usage{
print_provenance(x)
}
\arguments{
\item{x}{Object with ontology_provenance attribute.}
}
\description{
Prints provenance information attached to evaluated concept results.
}
\keyword{internal}
